future: fix future-state movement compilation on gcc 12.1.1
Our implementation of the idiom

        promise<T> pr;
        return pr.get_future();

Creates an uninitialized future_state in the promise object, and then moves
it into a future object.

When T is not trivial, our move implementation checks whether the
future_state has a result before moving it. But when T *is* trivial,
our implementation just copies the result's bytes - not checking whether
it was ever initialized.

Gcc 12 can detect in some cases that uninitialized data is being copied,
and warn about it ("uninitialized") - and Seastar fails to build on
some tests using the above idiom.

Because this uninitialized copy of raw bytes is intended, let's tell the
compiler not to warn about it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>